+++ Source/Modules/IOModules/Video/InputVideo/InputVideo.h
@@ -71,7 +71,7 @@ public:
     int				size_x;
     int				size_y;
     
-    int             frameRate;
+    float           frameRate;
     int             id;
     
     float *			intensity;
